Providing AbstractFunctionSectionSource for clarity
Also, avoids alias issue with CLASS
