+++ officefloor/polyglot/officepolyglot_kotlin/src/main/java/net/officefloor/polyglot/kotlin/KotlinFunctionSectionSource.java
@@ -26,14 +26,14 @@ import net.officefloor.compile.spi.section.SectionFunctionNamespace;
 import net.officefloor.compile.spi.section.source.SectionSource;
 import net.officefloor.compile.spi.section.source.SectionSourceContext;
 import net.officefloor.frame.api.function.ManagedFunction;
-import net.officefloor.plugin.section.clazz.ClassSectionSource;
+import net.officefloor.plugin.section.clazz.AbstractFunctionSectionSource;
 
 /**
  * Kotlin function {@link SectionSource}.
  * 
  * @author Daniel Sagenschneider
  */
-public class KotlinFunctionSectionSource extends ClassSectionSource {
+public class KotlinFunctionSectionSource extends AbstractFunctionSectionSource {
 
 	/**
 	 * {@link Property} name of the Kotlin function to use.
